+++ CHANGELOG.md
@@ -1,6 +1,6 @@
-# [2.0.1]()
-* `@Optional` annotation
-* Custom validation via default methods in config interfaces
+# [2.0.1](http://central.maven.org/maven2/com/github/tonybaines/gestalt/2.0.1) (TBA)
+* `@Optional` annotation for properties which need not be defined
+* Issue #5: Custom validation (via default methods in config interfaces)
 
 # [2.0.0](http://central.maven.org/maven2/com/github/tonybaines/gestalt/2.0.0) (2016-03-29)
 * *BREAKING CHANGE*: Recompiled with Java 8